web/requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
Pillow
pdf2image
unidecode
requests
rq
redis
